% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilis.R
\name{get_nps_code}
\alias{get_nps_code}
\title{get unicode and full names of national park units}
\usage{
get_nps_code()
}
\description{
get unicode and full names of national park units
}
\examples{
nps_code = get_nps_code()
}
