% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilis.R
\name{get_one_park}
\alias{get_one_park}
\title{get species list for one park from https://irma.nps.gov/NPSpecies/Search/SpeciesList}
\usage{
get_one_park(parkcode = "ACAD", categcode = 11)
}
\arguments{
\item{parkcode}{four digits code for parks from `rnpsp::get_nps_code()$UnitCode`.}

\item{categcode}{taxa groups code from `rnpsp::get_taxa_code()$Id`.
   Id                  Name
1   1               Mammals
2   2                 Birds
3   4              Reptiles
4   5            Amphibians
5   3                  Fish
6  11       Vascular Plants
7  12   Non-vascular Plants
8   7  Crab/Lobsters/Shrimp
9   9          Slugs/Snails
10  6     Spiders/Scorpions
11  8               Insects
12 10 Other Non-vertebrates
13 13                 Fungi
14 14              Protozoa
15 16             Chromista
16 15              Bacteria
17 17               Archaea}
}
\description{
get species list for one park from https://irma.nps.gov/NPSpecies/Search/SpeciesList
}
\examples{
get_one_park("ABLI", 11)
}
