% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nps_plants}
\alias{nps_plants}
\title{National Park Units' plant lists}
\format{A data frame with variables such as: \code{SciName}, \code{CommonNames},
  \code{Occurrence}, \code{Nativeness}, \code{Abundance} etc..
  Nativeness meta data available at \url{https://irma.nps.gov/NPSpecies/Lookup/GetNativenessMetadata};
  occurrence meta data available at \url{https://irma.nps.gov/NPSpecies/Lookup/GetOccurrenceMetadata};
  abundance meta data available at \url{https://irma.nps.gov/NPSpecies/Lookup/GetAbundanceMetadata};
  NPS tag meta data available at \url{https://irma.nps.gov/NPSpecies/Lookup/GetNPSTagsMetadata};
  The number of columns depend on the search level you defined.}
\usage{
nps_plants
}
\description{
Species list of vascular plants for all available 304 NPS units.
}
\keyword{datasets}
