% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nps_birds}
\alias{nps_birds}
\title{National Park Units' bird lists}
\format{A data frame with 23 variables, including: \code{SciName}, \code{CommonNames},
  \code{Occurrence}, \code{Nativeness}, \code{Abundance} etc..}
\usage{
nps_birds
}
\description{
Species list of birds for all available 308 NPS units.
}
\keyword{datasets}
