% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nps_plants}
\alias{nps_plants}
\title{National Park Units' plant lists}
\format{A data frame with 23 variables, including: \code{SciName}, \code{CommonNames},
  \code{Occurrence}, \code{Nativeness}, \code{Abundance} etc..}
\usage{
nps_plants
}
\description{
Species list of vascular plants for all available 304 NPS units.
}
\keyword{datasets}
